Rename workshop_table to WorkshopTable
